# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/companies": {
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "bizRegNo": {
                    "type": "string"
                  },
                  "representative": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessType": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessItem": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 30
                  }
                },
                "required": [
                  "name",
                  "bizRegNo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "만든 회사로 바로 전환된다",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "company": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "bizRegNo": {
                          "type": "string"
                        },
                        "representative": {
                          "nullable": true,
                          "type": "string"
                        },
                        "businessType": {
                          "nullable": true,
                          "type": "string"
                        },
                        "businessItem": {
                          "nullable": true,
                          "type": "string"
                        },
                        "address": {
                          "nullable": true,
                          "type": "string"
                        },
                        "phone": {
                          "nullable": true,
                          "type": "string"
                        },
                        "fiscalYearStartMonth": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "bizRegNo",
                        "representative",
                        "businessType",
                        "businessItem",
                        "address",
                        "phone",
                        "fiscalYearStartMonth"
                      ],
                      "additionalProperties": false
                    },
                    "accessToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "company"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "companies"
        ]
      }
    },
    "/api/companies/current": {
      "get": {
        "operationId": "CompaniesController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "bizRegNo": {
                      "type": "string"
                    },
                    "representative": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessType": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessItem": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "string"
                    },
                    "phone": {
                      "nullable": true,
                      "type": "string"
                    },
                    "fiscalYearStartMonth": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "bizRegNo",
                    "representative",
                    "businessType",
                    "businessItem",
                    "address",
                    "phone",
                    "fiscalYearStartMonth"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "권한 필요: settings.company (read)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "bizRegNo": {
                    "type": "string"
                  },
                  "representative": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessType": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessItem": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 100
                  },
                  "address": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 200
                  },
                  "phone": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 30
                  },
                  "fiscalYearStartMonth": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "bizRegNo": {
                      "type": "string"
                    },
                    "representative": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessType": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessItem": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "string"
                    },
                    "phone": {
                      "nullable": true,
                      "type": "string"
                    },
                    "fiscalYearStartMonth": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "bizRegNo",
                    "representative",
                    "businessType",
                    "businessItem",
                    "address",
                    "phone",
                    "fiscalYearStartMonth"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "권한 필요: settings.company (write)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/business-places": {
      "get": {
        "operationId": "CompaniesController_listPlaces",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "name": {
                        "type": "string"
                      },
                      "bizRegNo": {
                        "type": "string"
                      },
                      "representative": {
                        "nullable": true,
                        "type": "string"
                      },
                      "businessType": {
                        "nullable": true,
                        "type": "string"
                      },
                      "businessItem": {
                        "nullable": true,
                        "type": "string"
                      },
                      "address": {
                        "nullable": true,
                        "type": "string"
                      },
                      "isHeadquarters": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "bizRegNo",
                      "representative",
                      "businessType",
                      "businessItem",
                      "address",
                      "isHeadquarters"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "403": {
            "description": "권한 필요: settings.company (read)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_createPlace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "bizRegNo": {
                    "type": "string"
                  },
                  "representative": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessType": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessItem": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 100
                  },
                  "address": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "name",
                  "bizRegNo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "bizRegNo": {
                      "type": "string"
                    },
                    "representative": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessType": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessItem": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "string"
                    },
                    "isHeadquarters": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "bizRegNo",
                    "representative",
                    "businessType",
                    "businessItem",
                    "address",
                    "isHeadquarters"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "권한 필요: settings.company (write)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/business-places/{id}": {
      "patch": {
        "operationId": "CompaniesController_updatePlace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "bizRegNo": {
                    "type": "string"
                  },
                  "representative": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessType": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 50
                  },
                  "businessItem": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 100
                  },
                  "address": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "bizRegNo": {
                      "type": "string"
                    },
                    "representative": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessType": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessItem": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "string"
                    },
                    "isHeadquarters": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "bizRegNo",
                    "representative",
                    "businessType",
                    "businessItem",
                    "address",
                    "isHeadquarters"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "권한 필요: settings.company (write)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_removePlace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "권한 필요: settings.company (write)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
...
    "/api/invitations/{id}": {
      "delete": {
        "operationId": "CompaniesController_revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "권한 필요: settings.users (write)"
          }
        },
        "security": [
          {
            "wb_at": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/invitations/lookup": {
      "get": {
        "operationId": "CompaniesController_lookupInvitation",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companyName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "accountant",
                        "approver",
                        "employee"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "pending",
                        "accepted",
                        "expired",
                        "revoked"
                      ]
                    }
                  },
                  "required": [
                    "companyName",
                    "email",
                    "role",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "companies"
        ]
      }
    },
    "/api/invitations/accept": {
      "post": {
        "operationId": "CompaniesController_acceptInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 200
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "companyId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "accessToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "companyId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "companies"
        ]
      }
    }
  },
  "info": {
    "title": "WellBuddy ERP API",
    "description": "자동화 회계 ERP API. 웹은 쿠키(wb_at), 모바일 앱은 Bearer 토큰으로 인증한다.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wb_at"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}